add validateToken for the rest of the endpoints and add headers for axios requests

+++ backend/server.js
@@ -17,7 +17,7 @@ initializeApp({
 });
 const firebaseDb = getFirestore();
 
-app.get("/working-hours", (req, res) => {
+app.get("/working-hours", validateToken, (req, res) => {
   db.all(
     "SELECT tuntikirjaus_employee.firstname, tuntikirjaus_employee.lastname, worklog_worklog.hours FROM tuntikirjaus_employee JOIN worklog_worklog ON tuntikirjaus_employee.id = worklog_worklog.employee_id",
     (err, rows) => {
@@ -30,45 +30,53 @@ app.get("/working-hours", (req, res) => {
   );
 });
 
-app.get("/monthly-working-hours/:endYear/:years_back", (req, res) => {
-  let endYear = req.params.endYear;
-  const yearsBack = req.params.years_back;
-  let startYear = Number(endYear) - Number(yearsBack);
-  endYear = endYear.toString();
-  startYear = startYear.toString();
-  db.all(
-    "SELECT CAST(strftime('%Y', date) AS INTEGER) AS year, CAST(strftime('%m', date) AS INTEGER) AS month, SUM(hours) AS total_hours FROM worklog_worklog WHERE (strftime('%Y', date) between ? AND ?) AND (deleted = 0) GROUP BY year, month ORDER BY year, month",
-    [startYear, endYear],
-    (err, rows) => {
-      if (err) {
-        res.status(400).json({ error: err.message });
-        return;
+app.get(
+  "/monthly-working-hours/:endYear/:years_back",
+  validateToken,
+  (req, res) => {
+    let endYear = req.params.endYear;
+    const yearsBack = req.params.years_back;
+    let startYear = Number(endYear) - Number(yearsBack);
+    endYear = endYear.toString();
+    startYear = startYear.toString();
+    db.all(
+      "SELECT CAST(strftime('%Y', date) AS INTEGER) AS year, CAST(strftime('%m', date) AS INTEGER) AS month, SUM(hours) AS total_hours FROM worklog_worklog WHERE (strftime('%Y', date) between ? AND ?) AND (deleted = 0) GROUP BY year, month ORDER BY year, month",
+      [startYear, endYear],
+      (err, rows) => {
+        if (err) {
+          res.status(400).json({ error: err.message });
+          return;
+        }
+        res.json({ data: rows });
       }
-      res.json({ data: rows });
-    }
-  );
-});
+    );
+  }
+);
 
-app.get("/billability-working-hours/:endYear/:years_back", (req, res) => {
-  let endYear = req.params.endYear;
-  const yearsBack = req.params.years_back;
-  let startYear = Number(endYear) - Number(yearsBack);
-  endYear = endYear.toString();
-  startYear = startYear.toString();
-  db.all(
-    "SELECT CAST(strftime('%Y', date) AS INTEGER) AS year, CAST(strftime('%m', date) AS INTEGER) AS month, SUM(hours) AS total_hours, billable FROM worklog_worklog WHERE (strftime('%Y', date) between ? AND ?) AND (deleted = 0) GROUP BY year, month, billable ORDER BY year, month",
-    [startYear, endYear],
-    (err, rows) => {
-      if (err) {
-        res.status(400).json({ error: err.message });
-        return;
+app.get(
+  "/billability-working-hours/:endYear/:years_back",
+  validateToken,
+  (req, res) => {
+    let endYear = req.params.endYear;
+    const yearsBack = req.params.years_back;
+    let startYear = Number(endYear) - Number(yearsBack);
+    endYear = endYear.toString();
+    startYear = startYear.toString();
+    db.all(
+      "SELECT CAST(strftime('%Y', date) AS INTEGER) AS year, CAST(strftime('%m', date) AS INTEGER) AS month, SUM(hours) AS total_hours, billable FROM worklog_worklog WHERE (strftime('%Y', date) between ? AND ?) AND (deleted = 0) GROUP BY year, month, billable ORDER BY year, month",
+      [startYear, endYear],
+      (err, rows) => {
+        if (err) {
+          res.status(400).json({ error: err.message });
+          return;
+        }
+        res.json({ data: rows });
       }
-      res.json({ data: rows });
-    }
-  );
-});
+    );
+  }
+);
 
-app.get("/salary", (req, res) => {
+app.get("/salary", validateToken, (req, res) => {
   db.all(
     "SELECT date FROM payday_payday WHERE deleted = 0 ORDER BY date DESC",
     (err, rows) => {
@@ -81,7 +89,7 @@ app.get("/salary", (req, res) => {
   );
 });
 
-app.get("/salary_report/:selected_date", (req, res) => {
+app.get("/salary_report/:selected_date", validateToken, (req, res) => {
   const selectedDate = req.params.selected_date;
   db.all(
     "SELECT tuntikirjaus_employee.lastname, tuntikirjaus_employee.firstname, workphase_workphase.report_name, SUM(worklog_worklog.worker_hours) AS hours FROM worklog_worklog JOIN payday_payday ON worklog_worklog.payday_id = payday_payday.id JOIN workphase_workphase ON worklog_worklog.workphase_id = workphase_workphase.id JOIN tuntikirjaus_employee ON worklog_worklog.employee_id = tuntikirjaus_employee.id WHERE payday_payday.date= ? GROUP BY tuntikirjaus_employee.id, workphase_workphase.report_name ORDER BY tuntikirjaus_employee.lastname, tuntikirjaus_employee.firstname",
@@ -114,7 +122,10 @@ async function addUser(email, password, lastname, firstname) {
   return { userRecord, res };
 }
 
-app.post("/users", (req, res) => {
+app.post("/users", validateToken, (req, res) => {
+  if (req.user.admin !== true) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
   const email = req.body.email;
   const password = req.body.password;
   const lastname = req.body.lastname;
@@ -194,14 +205,20 @@ async function deleteDatabaseUser(id) {
   return res;
 }
 
-app.delete("/users/:id", (req, res) => {
+app.delete("/users/:id", validateToken, (req, res) => {
+  if (req.user.admin !== true) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
   const id = req.params.id;
   deleteAuthUser(id);
   deleteDatabaseUser(id);
   res.json(id);
 });
 
-app.get("/users/:id", async (req, res) => {
+app.get("/users/:id", validateToken, async (req, res) => {
+  if (req.user.admin !== true) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
   const id = req.params.id;
 
   async function getNames(id) {
@@ -229,7 +246,10 @@ app.get("/users/:id", async (req, res) => {
   res.json(data);
 });
 
-app.post("/users/:id", (req, res) => {
+app.post("/users/:id", validateToken, (req, res) => {
+  if (req.user.admin !== true) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
   const uid = req.params.id;
   const email = req.body.email;
   const lastname = req.body.lastname;
@@ -261,7 +281,10 @@ app.post("/users/:id", (req, res) => {
     });
 });
 
-app.post("/users/:id/change-password", (req, res) => {
+app.post("/users/:id/change-password", validateToken, (req, res) => {
+  if (req.user.admin !== true) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
   const uid = req.params.id;
   const password = req.body.password;
